[RPC] Add option to make rref.get_type not block.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50977

Adds a `blocking` flag that can be set to False to make this API return a `Future` to the type. This is to make this function non-blocking, mostly for a future change that will allow `rref.rpc_async()` to be completely non-blocking (it currently calls and waits for this function that issues an RPC in-line).
ghstack-source-id: 120602985

Differential Revision: [D25944582](https://our.internmc.facebook.com/intern/diff/D25944582/)