[PyTorch] Debug-gate static_assert in KernelFunction::makeFromUnboxedFunctor
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51367

Templight said that this assertion was taking about 5% of build time for RegisterCPU.cpp (a hopefully-representative example I picked to shorten my iteration cycle).

I've debug-gated it on the grounds that 1) we at least try to build
everything in debug mode and 2) optimized builds presumably take
longer in general, so we can more afford to pay the build time cost in
debug builds.

The win is not entirely clear; please see the test plan for details.
ghstack-source-id: 120697489

Differential Revision: [D26153793](https://our.internmc.facebook.com/intern/diff/D26153793/)