[QNNPACK, Sparsity] Sparse kernel with 4x8 blocking
Summary:
Larger blocking across M dim such as 8 in previous PR is likely
introducing wasted compute on the shapes being benchmarked.
Here we introduced 4x8 blocking of mrxnr. This helps 1) in packing
smaller data for small values of M and 2) for compute kernel it writes
same number of bytes but more contiguously. It is not certain but it
likely helps.

Test Plan:
q8gemm-sparse-test
fully-connected-sparse-test

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 04ab9880e5a763622f5bbafbe1d61be12b32dcff
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50590